requirements.txt
pyautogui>=0.9.54
keyboard>=0.13.5
pyperclip>=1.8.2
psutil>=5.9.6
pystray>=0.19.4
Pillow>=9.0.0 